requirements.txt
requests
beautifulsoup4
pyodbc
pandas
sqlalchemy
scikit-learn
python-dotenv
jupyter
matplotlib
seaborn
ipykernel